kmeeth/vecfc-absorption (#41)
* Added vecfc package to vecengine folder.

* Merged former vecfc contents into vecengine contents, removing unnecessary interfaces and clutter.

* Moved vecflushable from vecengine as it is fundamentally different and independent from it.

* Added vecfc package to vecengine folder.

* Merged former vecfc contents into vecengine contents, removing unnecessary interfaces and clutter.

* Moved vecflushable from vecengine as it is fundamentally different and independent from it.

* Reinstated interfaces with some changes, as they are still needed because of Sonic.

* Small additional tweaks.

* Fixed capitalization oversight from merging.

* Made callbacks public.

* Changed the way callbacks are handled to make them more easily integrable with Sonic.
