+++ utils/adapters/vector_to_dagidx.go
@@ -4,15 +4,15 @@ import (
 	"github.com/0xsoniclabs/consensus/abft/dagidx"
 	"github.com/0xsoniclabs/consensus/hash"
 	"github.com/0xsoniclabs/consensus/inter/idx"
-	"github.com/0xsoniclabs/consensus/vecfc"
+	"github.com/0xsoniclabs/consensus/vecengine"
 )
 
 type VectorSeqToDagIndexSeq struct {
-	*vecfc.HighestBeforeSeq
+	*vecengine.HighestBeforeSeq
 }
 
 type BranchSeq struct {
-	vecfc.BranchSeq
+	vecengine.BranchSeq
 }
 
 // Seq is a maximum observed e.Seq in the branch
@@ -32,9 +32,9 @@ func (b VectorSeqToDagIndexSeq) Get(i idx.Validator) dagidx.Seq {
 }
 
 type VectorToDagIndexer struct {
-	*vecfc.Index
+	*vecengine.Engine
 }
 
 func (v *VectorToDagIndexer) GetMergedHighestBefore(id hash.Event) dagidx.HighestBeforeSeq {
-	return VectorSeqToDagIndexSeq{v.Index.GetMergedHighestBefore(id)}
+	return VectorSeqToDagIndexSeq{v.Engine.GetMergedHighestBefore(id).(*vecengine.HighestBeforeSeq)}
 }
