+++ vecengine/store_vectors.go
@@ -1,11 +1,11 @@
-package vecfc
+package vecengine
 
 import (
 	"github.com/0xsoniclabs/consensus/hash"
 	"github.com/0xsoniclabs/consensus/kvdb"
 )
 
-func (vi *Index) getBytes(table kvdb.Store, id hash.Event) []byte {
+func (vi *Engine) getBytes(table kvdb.Store, id hash.Event) []byte {
 	key := id.Bytes()
 	b, err := table.Get(key)
 	if err != nil {
@@ -14,7 +14,7 @@ func (vi *Index) getBytes(table kvdb.Store, id hash.Event) []byte {
 	return b
 }
 
-func (vi *Index) setBytes(table kvdb.Store, id hash.Event, b []byte) {
+func (vi *Engine) setBytes(table kvdb.Store, id hash.Event, b []byte) {
 	key := id.Bytes()
 	err := table.Put(key, b)
 	if err != nil {
@@ -23,7 +23,7 @@ func (vi *Index) setBytes(table kvdb.Store, id hash.Event, b []byte) {
 }
 
 // GetLowestAfter reads the vector from DB
-func (vi *Index) GetLowestAfter(id hash.Event) *LowestAfterSeq {
+func (vi *Engine) GetLowestAfter(id hash.Event) *LowestAfterSeq {
 	if bVal, okGet := vi.cache.LowestAfterSeq.Get(id); okGet {
 		return bVal.(*LowestAfterSeq)
 	}
@@ -37,7 +37,7 @@ func (vi *Index) GetLowestAfter(id hash.Event) *LowestAfterSeq {
 }
 
 // GetHighestBefore reads the vector from DB
-func (vi *Index) GetHighestBefore(id hash.Event) *HighestBeforeSeq {
+func (vi *Engine) GetHighestBefore(id hash.Event) *HighestBeforeSeq {
 	if bVal, okGet := vi.cache.HighestBeforeSeq.Get(id); okGet {
 		return bVal.(*HighestBeforeSeq)
 	}
@@ -51,14 +51,14 @@ func (vi *Index) GetHighestBefore(id hash.Event) *HighestBeforeSeq {
 }
 
 // SetLowestAfter stores the vector into DB
-func (vi *Index) SetLowestAfter(id hash.Event, seq *LowestAfterSeq) {
+func (vi *Engine) SetLowestAfter(id hash.Event, seq *LowestAfterSeq) {
 	vi.setBytes(vi.table.LowestAfterSeq, id, *seq)
 
 	vi.cache.LowestAfterSeq.Add(id, seq, uint(len(*seq)))
 }
 
 // SetHighestBefore stores the vectors into DB
-func (vi *Index) SetHighestBefore(id hash.Event, seq *HighestBeforeSeq) {
+func (vi *Engine) SetHighestBefore(id hash.Event, seq *HighestBeforeSeq) {
 	vi.setBytes(vi.table.HighestBeforeSeq, id, *seq)
 
 	vi.cache.HighestBeforeSeq.Add(id, seq, uint(len(*seq)))
