+++ vecflushable/vecflushable.go
