+++ vecflushable/vecflushable_test.go
